move bladectl config into correct directory
